feat: serviceテストの追加
